backend/requirements.txt
aiofiles
aiohttp==3.8.4
aiosignal
annotated-types
annoy==1.17.3
anyio==3.7.1
argilla
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
chromadb==0.3.26
click==8.1.4
clickhouse-connect==0.6.6
cohere==4.11.2
coloredlogs==15.0.1
commonmark==0.9.1
cryptography==41.0.1
dataclasses-json==0.5.9
Deprecated==1.2.14
dnspython==2.3.0
duckdb==0.8.1
duckduckgo-search==3.8.3
et-xmlfile==1.1.0
exceptiongroup==1.1.2
fastapi==0.100.0
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.1
google-auth==2.21.0
google-cloud-aiplatform==1.27.1
google-cloud-bigquery==3.11.3
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.56.0
grpcio-status==1.56.0
h11==0.14.0
h2==4.1.0
hnswlib==0.7.0
hpack==4.0.0
httpcore==0.16.3
httptools==0.6.0
httpx==0.23.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
joblib==1.3.1
langchain==0.0.228
langchainplus-sdk==0.0.20
lxml==4.9.3
lz4==4.3.2
Markdown==3.4.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==10.0.0
posthog==3.0.1
proto-plus==1.22.3
protobuf==4.23.4
pulsar-client==3.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
pydantic_core==2.1.2
Pygments==2.15.1
pymongo==4.4.0
pypandoc==1.11
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
socksio==1.0.0
SQLAlchemy==2.0.18
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.65.0
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.7.1
unstructured==0.8.0
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.16.0
zstandard==0.21.0
